checkpoint - migrate to using long/Long to represent any integer ID value (rowId, objectId, etc). (#233)

+++ ms2extensions/src/org/labkey/ms2extensions/PeptideCountPepXmlDataHandler.java
@@ -90,7 +90,7 @@ public class PeptideCountPepXmlDataHandler implements ExperimentDataHandler
     }
 
     @Override
-    public void runMoved(ExpData newData, Container container, Container targetContainer, String oldRunLSID, String newRunLSID, User user, int oldDataRowID) throws ExperimentException
+    public void runMoved(ExpData newData, Container container, Container targetContainer, String oldRunLSID, String newRunLSID, User user, long oldDataRowID) throws ExperimentException
     {
         _realHandler.runMoved(newData, container, targetContainer, oldRunLSID, newRunLSID, user, oldDataRowID);
     }
